Merge remote-tracking branch 'upstream/master'
There were merge conflicts due to my previous elimination of a local rubric from the Te Deum. Since that rubric exists in all the versions except the three from which I had eliminated it, I have restored it in Latin, English and Italian. However, I rephrased the Latin to make it clearer that this is a local custom.